+++ app/assets/javascripts/analyses/index.js
@@ -6,6 +6,10 @@ $(function() {
 
   $('#nav-link-analysis').addClass('active');
 
+  document.querySelectorAll('[data-bs-toggle="tooltip"]').forEach(function (element) {
+    new bs.Tooltip(element);
+  });
+
   $('#new_analysis').on('ajax:success', function(event) {
     showDialog('dialog-execute');
   });