+++ app/assets/javascripts/channels/evaluations/manage.js
@@ -14,6 +14,7 @@ consumer.subscriptions.create('EvaluationChannel', {
           column.attr('title', '結果を確認');
           $(`${trId} > td[class*=performed_at]`).text(evaluation.performed_at);
           $(`${trId} > td[class*=state]`).text('0%完了');
+          new bs.Tooltip(column);
           break;
         case 'completed':
           column.removeClass('table-warning');
@@ -27,15 +28,28 @@ consumer.subscriptions.create('EvaluationChannel', {
           column.removeClass('table-warning cursor-pointer');
           column.addClass('table-danger cursor-auto');
           column.attr('data-state', evaluation.state);
+          column.attr('title', '');
+          column.attr('data-bs-original-title', '');
           $(`${trId} > td[class*=state]`).text('エラー');
+          bs.Tooltip.getInstance(column).dispose();
           break;
         default:
           this.updateProgress(trId, evaluation);
       }
     } else {
+      let elements = document.querySelectorAll('[data-bs-toggle="tooltip"]');
+      elements.forEach(function (element) {
+        bs.Tooltip.getInstance(element).dispose();
+      });
+
       $.ajax({
         url: location.href,
         dataType: 'script',
+      }).done(function() {
+        elements = document.querySelectorAll('[data-bs-toggle="tooltip"]');
+        elements.forEach(function (element) {
+          new bs.Tooltip(element);
+        });
       });
     }
   },
@@ -46,13 +60,18 @@ consumer.subscriptions.create('EvaluationChannel', {
     }
 
     const href = `/alterf/evaluations/${evaluation.evaluation_id}/download`;
-    $(`${trId} > td.download`).append(
+    const button = $(
       `<a data-remote='true' href='${href}'>` +
-        '<button class="btn btn-success">' +
+        '<button class="btn btn-success" title="評価レースをダウンロード" ' +
+          'data-bs-toggle="tooltip" data-bs-trigger="hover">' +
           '<span class="bi bi-download"></span>' +
         '</button>' +
       '</a>'
     );
+    $(`${trId} > td.download`).append(button);
+    button.ready(function() {
+      new bs.Tooltip($(`${trId} > td.download > a > button`));
+    });
   },
 
   updateProgress(trId, evaluation) {