+++ app/assets/packs/evaluation/races/show.js
@@ -1,0 +1,2 @@
+import '../../../stylesheets/evaluation/races/show';
+import '../../../javascripts/evaluation/races/show';