+++ app/channels/evaluation/race_channel.rb
@@ -1,0 +1,5 @@
+class Evaluation::RaceChannel < ApplicationCable::Channel
+  def subscribed
+    stream_from 'evaluation_race'
+  end
+end