+++ app/helpers/evaluation_helper.rb
@@ -15,16 +15,6 @@ module EvaluationHelper
     ]
   end
 
-  def collapse_params(id, expanded)
-    {
-      type: 'button',
-      'data-bs-toggle' => 'collapse',
-      'data-bs-target' => "##{id}",
-      'aria-controls' => id,
-      'aria-expanded' => expanded,
-    }
-  end
-
   def progress(evaluation)
     case evaluation.state
     when 'completed'
@@ -34,13 +24,13 @@ module EvaluationHelper
     when 'waiting'
       '実行待ち'
     else
-      if evaluation.data.empty?
+      if evaluation.races.empty?
         '0%完了'
       else
-        completed_data_size = evaluation.data.to_a.count do |datum|
-          datum.prediction_results.present?
+        completed_data_size = evaluation.races.to_a.count do |race|
+          race.test_data.present?
         end
-        "#{(100 * completed_data_size / evaluation.data.size.to_f).round(0)}%完了"
+        "#{(100 * completed_data_size / evaluation.races.size.to_f).round(0)}%完了"
       end
     end
   end
@@ -62,10 +52,18 @@ module EvaluationHelper
     %w[processing completed].include?(state) ? '結果を確認' : ''
   end
 
-  def datum_row_class(numbers, datum)
-    return 'warning' if datum.prediction_results.empty?
+  def race_row_class(numbers, race)
+    return 'warning' if race.test_data.empty?
 
-    numbers.include?(datum.ground_truth) ? 'success' : 'danger'
+    numbers.include?(race.ground_truth) ? 'success' : 'danger'
+  end
+
+  def cell_class(prediction_result, feature_name)
+    if feature_name == 'order'
+      'table-dark'
+    else
+      prediction_result ? 'table-success' : 'table-danger'
+    end
   end
 
   def span_color(number, ground_truth)
@@ -86,7 +84,13 @@ module EvaluationHelper
     return unless evaluation.state == 'completed'
 
     link_to(evaluation_download_path(evaluation.evaluation_id), remote: true) do
-      tag.button(class: 'btn btn-success') do
+      attribute = {
+        class: 'btn btn-success',
+        title: '評価レースをダウンロード',
+        data: {'bs-toggle' => 'tooltip'},
+      }
+
+      tag.button(attribute) do
         tag.span(class: 'bi bi-download')
       end
     end