+++ app/models/denebola/entry.rb
@@ -1,0 +1,3 @@
+class Denebola::Entry < Denebola::Base
+  belongs_to :race
+end