+++ spec/models/evaluation/race_spec.rb
@@ -2,7 +2,7 @@
 
 require 'rails_helper'
 
-describe Evaluation::Datum, type: :model do
+describe Evaluation::Race, type: :model do
   default_attribute = {
     evaluation_id: 1,
     race_id: '0',
@@ -11,16 +11,16 @@ describe Evaluation::Datum, type: :model do
     ground_truth: 1,
   }
 
-  shared_context '評価データ情報を作成する' do |attribute: default_attribute|
+  shared_context '評価レース情報を作成する' do |attribute: default_attribute|
     before do
       @evaluation ||= create(:evaluation)
-      @evaluation_datum = @evaluation.data.create!(attribute)
+      @evaluation_race = @evaluation.races.create!(attribute)
     end
   end
 
   shared_examples '結果をインポートすると例外が発生すること' do |file, e|
     it_is_asserted_by do
-      @evaluation_datum.import_prediction_results(file)
+      @evaluation_race.import_prediction_results!(file)
     rescue e
       true
     end
@@ -37,7 +37,7 @@ describe Evaluation::Datum, type: :model do
 
       CommonHelper.generate_test_case(valid_attribute).each do |attribute|
         context "#{attribute}を指定した場合" do
-          before(:all) { @object = build(:datum, attribute) }
+          before(:all) { @object = build(:evaluation_race, attribute) }
 
           it_behaves_like 'バリデーションエラーにならないこと'
         end
@@ -53,7 +53,7 @@ describe Evaluation::Datum, type: :model do
 
           before(:all) do
             attribute = absent_keys.map {|key| [key, nil] }.to_h
-            @object = build(:datum, attribute)
+            @object = build(:evaluation_race, attribute)
             @object.validate
           end
 
@@ -71,7 +71,7 @@ describe Evaluation::Datum, type: :model do
           expected_error = attribute.keys.map {|key| [key, 'invalid_parameter'] }.to_h
 
           before(:all) do
-            @object = build(:datum, attribute)
+            @object = build(:evaluation_race, attribute)
             @object.validate
           end
 
@@ -84,32 +84,26 @@ describe Evaluation::Datum, type: :model do
   describe '#destroy' do
     describe '正常系' do
       include_context 'トランザクション作成'
-      include_context '評価データ情報を作成する'
-      before { @other_evaluation_datum = @evaluation_datum }
-      include_context '評価データ情報を作成する'
+      include_context '評価レース情報を作成する'
+      before { @other_evaluation_race = @evaluation_race }
+      include_context '評価レース情報を作成する'
       before do
-        @evaluation_datum.prediction_results.create!(number: 1)
-        @other_evaluation_datum.prediction_results.create!(number: 1)
-        @evaluation_datum.destroy
+        @evaluation_race.test_data.create!(number: 1)
+        @other_evaluation_race.test_data.create!(number: 1)
+        @evaluation_race.destroy
       end
 
-      it '紐づいているPrediction::Resultが削除されていること' do
+      it '紐づいているEvaluation::Race::TestDatumが削除されていること' do
         is_asserted_by do
-          query = {
-            predictable_id: @evaluation_datum.id,
-            predictable_type: 'Evaluation::Datum',
-          }
-          not Prediction::Result.exists?(query)
+          query = {evaluation_race_id: @evaluation_race.id}
+          not Evaluation::Race::TestDatum.exists?(query)
         end
       end
 
       it '紐づいていないPrediction::Resultが削除されていないこと' do
         is_asserted_by do
-          query = {
-            predictable_id: @other_evaluation_datum.id,
-            predictable_type: 'Evaluation::Datum',
-          }
-          Prediction::Result.exists?(query)
+          query = {evaluation_race_id: @other_evaluation_race.id}
+          Evaluation::Race::TestDatum.exists?(query)
         end
       end
     end
@@ -120,20 +114,23 @@ describe Evaluation::Datum, type: :model do
       file = Rails.root.join('spec/fixtures/prediction.yml')
       include_context 'トランザクション作成'
       include_context 'ActionCableのモックを作成'
-      include_context '評価データ情報を作成する'
+      include_context '評価レース情報を作成する'
       before do
         @called = false
-        @evaluation_datum.import_prediction_results(file)
+        @evaluation_race.import_prediction_results!(file)
       end
 
       it '予測結果情報が登録されていること' do
         won = [3, 5, 11, 17]
-        results = @evaluation_datum.prediction_results
-
-        is_asserted_by { results.where(won: true).pluck(:number).sort == won }
+        lose = (1..18).to_a - won
+        test_data = @evaluation_race.test_data
 
         is_asserted_by do
-          results.where(won: false).pluck(:number).sort == (1..18).to_a - won
+          test_data.where(prediction_result: true).pluck(:number).sort == won
+        end
+
+        is_asserted_by do
+          test_data.where(prediction_result: false).pluck(:number).sort == lose
         end
       end
 
@@ -146,7 +143,7 @@ describe Evaluation::Datum, type: :model do
       context 'ファイルが存在しない場合' do
         file = Rails.root.join('spec/fixtures/not_exist.yml')
         include_context 'トランザクション作成'
-        include_context '評価データ情報を作成する'
+        include_context '評価レース情報を作成する'
 
         it_behaves_like '結果をインポートすると例外が発生すること', file, Errno::ENOENT
       end
@@ -155,7 +152,7 @@ describe Evaluation::Datum, type: :model do
         context '配列の場合' do
           file = Rails.root.join('spec/fixtures/array.yml')
           include_context 'トランザクション作成'
-          include_context '評価データ情報を作成する'
+          include_context '評価レース情報を作成する'
           before(:all) { File.open(file, 'w') {|f| YAML.dump([3, 5, 11, 17], f) } }
           after(:all) { FileUtils.rm(file) }
 
@@ -166,7 +163,7 @@ describe Evaluation::Datum, type: :model do
         context 'ハッシュの値が数値でない場合' do
           file = Rails.root.join('spec/fixtures/invalid_value.yml')
           include_context 'トランザクション作成'
-          include_context '評価データ情報を作成する'
+          include_context '評価レース情報を作成する'
           before(:all) { File.open(file, 'w') {|f| YAML.dump({'invalid' => 1}, f) } }
           after(:all) { FileUtils.rm(file) }